functions/generate_report/requirements.txt
functions-framework==3.*
claude-agent-sdk>=0.1.79
anthropic>=0.52
google-cloud-firestore>=2.16
google-cloud-storage>=2.18
resend>=2.0
markdown>=3.6
